Move impl ToTokens for Entity and Schema
